RUB-243: extract chainstate_io.go — IO/genesis/utxo helpers (696→314 lines)
Pure block extraction, zero logic changes. chainstate.go loses 382 lines (-61 complexity), chainstate_io.go gains 378 lines (+61 complexity), net 0.

Codacy annotations:
- L130 (0o700 on os.MkdirAll): directory permissions require execute bit for traverse; 0o700 is correct, 0o600 would block directory access
- L411 (os.OpenFile tmpPath): path derived from os.Getpid + atomic counter, no external/user input; false positive

parity-exempt: code organization only, no Rust counterpart.
